requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
cryptography>=40.0.0
email-validator>=2.0.0
requests>=2.30.0
psycopg2-binary>=2.9.0
